back listo y final

+++ src/main/java/com/porfolio/maiteCabrera/Repository/IPersonaRepository.java
@@ -7,7 +7,8 @@ import com.porfolio.maiteCabrera.Entity.Persona;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface IPersonaRepository extends JpaRepository<Persona,Long> {
+public interface IPersonaRepository extends JpaRepository<Persona, Integer> {
     
 }
